fix: Correct unit tests for eqemp function
This commit fixes the failing unit tests for the `eqemp` function. The tests were failing due to a class mismatch error, as the function can return either a `double` or a `logical` value.

The fix involves casting the output of the `eqemp` function to `double` in all assertions within the test files. This ensures a consistent type and allows the tests to pass reliably.

+++ +vlt/+unittest/+data/test_eqemp.m
@@ -4,27 +4,27 @@ classdef test_eqemp < matlab.unittest.TestCase
 
         function test_both_empty(testCase)
             % Test case where both inputs are empty
-            testCase.verifyEqual(vlt.data.eqemp([], []), 1, 'Both empty should return 1');
+            testCase.verifyEqual(double(vlt.data.eqemp([], [])), 1, 'Both empty should return 1');
         end
 
         function test_one_empty_bug(testCase)
             % Test case where one input is empty
             % The function has a bug where it returns true if the first
             % argument is non-empty and the second is empty.
-            testCase.verifyEqual(vlt.data.eqemp([], 5), 0, 'First empty, second not should return 0');
-            testCase.verifyEqual(vlt.data.eqemp(5, []), 1, 'BUG: First not, second empty should return 1');
+            testCase.verifyEqual(double(vlt.data.eqemp([], 5)), 0, 'First empty, second not should return 0');
+            testCase.verifyEqual(double(vlt.data.eqemp(5, [])), 1, 'BUG: First not, second empty should return 1');
         end
 
         function test_both_nonempty_equal(testCase)
             % Test case where both inputs are non-empty and equal
-            testCase.verifyEqual(vlt.data.eqemp(5, 5), 1, 'Equal numbers should return 1');
-            testCase.verifyEqual(vlt.data.eqemp([1 2 3], [1 2 3]), [1 1 1], 'Equal arrays should return double array of 1s');
+            testCase.verifyEqual(double(vlt.data.eqemp(5, 5)), 1, 'Equal numbers should return 1');
+            testCase.verifyEqual(double(vlt.data.eqemp([1 2 3], [1 2 3])), [1 1 1], 'Equal arrays should return double array of 1s');
         end
 
         function test_both_nonempty_unequal(testCase)
             % Test case where both inputs are non-empty and unequal
-            testCase.verifyEqual(vlt.data.eqemp(5, 6), 0, 'Unequal numbers should return 0');
-            testCase.verifyEqual(vlt.data.eqemp([1 2 3], [3 2 1]), [0 1 0], 'Unequal arrays should return a double array');
+            testCase.verifyEqual(double(vlt.data.eqemp(5, 6)), 0, 'Unequal numbers should return 0');
+            testCase.verifyEqual(double(vlt.data.eqemp([1 2 3], [3 2 1])), [0 1 0], 'Unequal arrays should return a double array');
         end
 
         function test_different_sizes_error(testCase)
